requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.32
psycopg[binary]==3.3.2
pydantic==1.10.13
pytest==8.3.2
pytest-cov==5.0.0
python-dotenv==1.0.1